requirements.txt
# Base
chromadb>=0.4.18
pydantic>=2.5.2
numpy>=1.26.2
pandas>=2.1.4

# Procesamiento de texto
nltk>=3.8.1
spacy>=3.7.2
python-magic>=0.4.27

# Utilidades
python-dotenv>=1.0.0
tqdm>=4.66.1
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
